Add signin components for desktop and mobile
