+++ frontend/app/[lang]/signin/components/index.ts
@@ -1,0 +1,2 @@
+export * from './Desktop'
+export * from './Mobile'
